requirements.txt
Django==1.6
Jinja2==2.7.1
MarkupSafe==0.18
Pillow==2.1.0
Pygments==1.6
South==0.8.2
Sphinx==1.2b3
balanced==0.11.15
bpython==0.12
certifi==0.0.8
chardet==1.0.1
distribute==0.7.3
django-allauth==0.13.0
django-appconf==0.6
django-autoslug==1.7.2
django-avatar==2.0
git+https://github.com/blorenz/django-balanced
django-braces==1.2.2
django-crispy-forms==1.3.2
django-csvimport==1.0
django-debug-toolbar==0.9.4
django-extensions==1.2.5
django-filebrowser==3.5.3
django-filter==0.7
django-floppyforms==1.1
django-grappelli==2.5.1
django-model-utils==1.5.0
django-reversion==1.8.0
django-social-auth==0.7.28
django-threadedcomments==0.9.0
django-tinymce==1.5.2
django-vanilla-views==1.0.1
django-widget-tweaks==1.3
djangorestframework==2.3.7
docutils==0.11
gunicorn==17.5
httplib2==0.8
ipython==1.1.0
iso8601==0.1.4
mock==0.8.0
nose==1.3.0
numpy==1.7.1
oauth2==1.5.211
oauthlib==0.6.0
psycopg2==2.5.1
python-openid==2.2.5
pyzmq==13.1.0
requests==1.2.3
requests-oauthlib==0.3.3
simplejson==2.3.2
six==1.3.0
tornado==3.1.1
wsgiref==0.1.2
easy-thumbnails==1.5
django-import-export
